add: idempotence, fix: alias no written

+++ environment.sh
@@ -5,7 +5,7 @@ set -euo pipefail
 echo "--- 1. Setting up termux configuration and dependencies ---"
 termux-setup-storage
 pkg update -y
-pkg install curl git -y
+pkg install curl git which -y
 
 echo "--- Installing Acode plugin ---"
 if curl -sL https://raw.githubusercontent.com/bajrangCoder/acode-plugin-acodex/main/installServer.sh | bash; then
@@ -16,22 +16,26 @@ else
 fi
 
 echo "--- Installing Termux Alpine ---"
+
 ALPINE_SCRIPT="TermuxAlpine.sh"
 ALPINE_URL="https://raw.githubusercontent.com/Hax4us/TermuxAlpine/master/TermuxAlpine.sh"
 
-if ! curl -LO "$ALPINE_URL"; then
-  echo "ERROR: TermuxAlpine.sh download failed " >&2
-  exit 1
-fi
+if which startalpine >/dev/null; then
+  echo "--- Termux Alpine is already installed"
+else
+  if ! curl -LO "$ALPINE_URL"; then
+    echo "ERROR: TermuxAlpine.sh download failed " >&2
+    exit 1
+  fi
 
-echo "Starting Alpine installation ..."
-if ! bash "$ALPINE_SCRIPT"; then
-  echo "ERROR: Alpine Terminux installation has failed." >&2
+  if ! bash "$ALPINE_SCRIPT"; then
+    echo "ERROR: Alpine Terminux installation has failed." >&2
+    rm -f "$ALPINE_SCRIPT"
+    exit 1
+  fi
   rm -f "$ALPINE_SCRIPT"
-  exit 1
+  echo "Termux Alpine installed."
 fi
-rm -f "$ALPINE_SCRIPT"
-echo "Termux Alpine installed."
 
 echo "--- Building and Installing the Pawn compiler (pawncc) ---"
 
@@ -47,6 +51,7 @@ apk add git cmake alpine-sdk linux-headers
 
 PAWN_DIR="/tmp/pawn-compiler"
 echo "Cloning compiler source
+rm -fr "\$PAWN_DIR"
 git clone https://github.com/openmultiplayer/compiler.git "\$PAWN_DIR"
 
 cd "\$PAWN_DIR/source/compiler"
@@ -61,10 +66,10 @@ mv pawncc /usr/bin/
 mv libpawnc.so /usr/lib/
 
 rm -fr "\$PAWN_DIR"
-echo "alias pawncc='pawncc -Dgamemodes -i../qawno/include -d3 -Z -\(+ -;+' >> ~/.profile"
+echo "alias pawncc='pawncc -Dgamemodes -i../qawno/include -d3 -Z -\(+ -;+' > ~/.profile"
 echo "alias pawncc-old='pawncc -Dgamemodes -i../pawno/include -d3 -Z -\(+ -;+' >> ~/.profile"
-echo "alias install-omp='wget https://github.com/openmultiplayer/open.mp/releases/download/v1.4.0.2779/open.mp-linux-x86.tar.gz && tar -xf open.mp-linux-x86.tar.gz && cp ./Server/omp-server . && rm -fr Server open.mp-linux-x86.tar.gz'"
-echo "alias install-samp='wget https://gta-multiplayer.cz/downloads/samp037svr_R2-2-1.tar.gz && tar -xf samp037svr_R2-2-1.tar.gz && cp samp03/samp03svr samp03/samp-npc . && rm -fr samp03 samp037svr_R2-2-1.tar.gz'"
+echo "alias install-omp='wget https://github.com/openmultiplayer/open.mp/releases/download/v1.4.0.2779/open.mp-linux-x86.tar.gz && tar -xf open.mp-linux-x86.tar.gz && cp ./Server/omp-server . && rm -fr Server open.mp-linux-x86.tar.gz' >> ~/.profile"
+echo "alias install-samp='wget https://gta-multiplayer.cz/downloads/samp037svr_R2-2-1.tar.gz && tar -xf samp037svr_R2-2-1.tar.gz && cp samp03/samp03svr samp03/samp-npc . && rm -fr samp03 samp037svr_R2-2-1.tar.gz' >> ~/.profile"
 source ~/.profile
 pawncc -v
 echo "Pawn compiler has been installed successfully in Alpine"
